Refactored schema i18n out of imports directory in attempt to make it behave correctly on production.
