+++ imports/api/data/reviews.js
@@ -92,37 +92,6 @@ Reviews.schema = new SimpleSchema(
 					}
 				}
 			},
-			/*
-			After working so hard to get this right,
-			here is why I have removed this custom validator:
-			There is no way to skip custom validation. However,
-			frontend wanted the insertion form to allow users
-			to skip this step under certain circumstances. This
-			would have been a cause for concern, except in the
-			other circumstances this field is auto-filled from
-			the collection and set to read-only. There thus
-			seems to be no use case for this check. But I worked
-			hard on it, so here it is in case I ever need to
-			refer back to its code, or even use it again in the
-			project, perhaps for testing.
-		*/
-			// custom: function() {
-			// 	if (Meteor.isClient && this.isSet) {
-			// 		Meteor.call("companies.doesCompanyExist", this.value, (error, result) => {
-			// 			if (!result) {
-			// 				this.validationContext.addValidationErrors([{
-			// 					name: "companyName",
-			// 					type: "noCompanyWithThatName",
-			// 				}]);
-			// 			}
-			// 		});
-			// 	}
-			// 	else if (Meteor.isServer && this.isSet) {
-			// 		if(!Companies.hasEntry(this.value)) {
-			// 			return "noCompanyWithThatName";
-			// 		}
-			// 	}
-			// },
 		},
 		companyId: {
 			type: String,
@@ -357,122 +326,6 @@ Reviews.schema = new SimpleSchema(
 	},
 	{ tracker: Tracker }
 );
-
-const reviewLabels = function() {
-	return {
-		_id: i18n.__("SimpleSchema.labels.Reviews._id", {
-			_locale: i18n.getLocale(),
-		}),
-		submittedBy: i18n.__("SimpleSchema.labels.Reviews.submittedBy", {
-			_locale: i18n.getLocale(),
-		}),
-		companyName: i18n.__("SimpleSchema.labels.Reviews.companyName", {
-			_locale: i18n.getLocale(),
-		}),
-		companyId: i18n.__("SimpleSchema.labels.Reviews.companyId", {
-			_locale: i18n.getLocale(),
-		}),
-		reviewTitle: i18n.__("SimpleSchema.labels.Reviews.reviewTitle", {
-			_locale: i18n.getLocale(),
-		}),
-		locations: i18n.__("SimpleSchema.labels.Reviews.locations", {
-			_locale: i18n.getLocale(),
-		}),
-		jobTitle: i18n.__("SimpleSchema.labels.Reviews.jobTitle", {
-			_locale: i18n.getLocale(),
-		}),
-		numberOfMonthsWorked: i18n.__(
-			"SimpleSchema.labels.Reviews.numberOfMonthsWorked",
-			{ _locale: i18n.getLocale() }
-		),
-		pros: i18n.__("SimpleSchema.labels.Reviews.pros", {
-			_locale: i18n.getLocale(),
-		}),
-		cons: i18n.__("SimpleSchema.labels.Reviews.cons", {
-			_locale: i18n.getLocale(),
-		}),
-		wouldRecommendToOtherJobSeekers: i18n.__(
-			"SimpleSchema.labels.Reviews.wouldRecommendToOtherJobSeekers",
-			{ _locale: i18n.getLocale() }
-		),
-		healthAndSafety: i18n.__(
-			"SimpleSchema.labels.Reviews.healthAndSafety",
-			{
-				_locale: i18n.getLocale(),
-			}
-		),
-		managerRelationship: i18n.__(
-			"SimpleSchema.labels.Reviews.managerRelationship",
-			{ _locale: i18n.getLocale() }
-		),
-		workEnvironment: i18n.__(
-			"SimpleSchema.labels.Reviews.workEnvironment",
-			{
-				_locale: i18n.getLocale(),
-			}
-		),
-		benefits: i18n.__("SimpleSchema.labels.Reviews.benefits", {
-			_locale: i18n.getLocale(),
-		}),
-		overallSatisfaction: i18n.__(
-			"SimpleSchema.labels.Reviews.overallSatisfaction",
-			{ _locale: i18n.getLocale() }
-		),
-		additionalComments: i18n.__(
-			"SimpleSchema.labels.Reviews.additionalComments",
-			{ _locale: i18n.getLocale() }
-		),
-		datePosted: i18n.__("SimpleSchema.labels.Reviews.datePosted", {
-			_locale: i18n.getLocale(),
-		}),
-		upvotes: i18n.__("SimpleSchema.labels.Reviews.upvotes", {
-			_locale: i18n.getLocale(),
-		}),
-		downvotes: i18n.__("SimpleSchema.labels.Reviews.downvotes", {
-			_locale: i18n.getLocale(),
-		}),
-	};
-};
-
-const reviewErrorMessages = function(locale) {
-	return {
-		needsFiveWords: i18n.__("SimpleSchema.custom.Reviews.needsFiveWords", {
-			_locale: locale,
-		}),
-		noCompanyWithThatName: i18n.__(
-			"SimpleSchema.custom.noCompanyWithThatName",
-			{ _locale: locale }
-		),
-		sessionError: i18n.__("SimpleSchema.custom.sessionError", {
-			_locale: locale,
-		}),
-	};
-};
-
-const englishReviews = reviewErrorMessages("en");
-if (Meteor.isClient && Meteor.isProduction) {
-	i18n.loadLocale("es");
-}
-const spanishReviews = reviewErrorMessages("es");
-
-console.log("english review error messages");
-console.log(englishReviews);
-console.log("spanish review error messages");
-console.log(spanishReviews);
-
-Reviews.schema.labels(reviewLabels());
-
-Reviews.schema.messageBox.messages({
-	en: englishReviews,
-	es: spanishReviews,
-});
-
-i18n.onChangeLocale(function(newLocale) {
-	// if (Meteor.isDevelopment) console.log("REVIEWS: " + newLocale);
-	console.log("REVIEWS: " + newLocale);
-	Reviews.schema.messageBox.setLanguage(newLocale);
-	Reviews.schema.labels(reviewLabels());
-});
 
 Reviews.attachSchema(Reviews.schema, { replace: true });
