+++ imports/api/data/salaries.js
@@ -132,66 +132,6 @@ Salaries.schema = new SimpleSchema(
 	{ tracker: Tracker }
 );
 
-const salaryLabels = function() {
-	return {
-		_id: i18n.__("SimpleSchema.labels.Salaries._id"),
-		submittedBy: i18n.__("SimpleSchema.labels.Salaries.submittedBy", {
-			_locale: i18n.getLocale(),
-		}),
-		companyName: i18n.__("SimpleSchema.labels.Salaries.companyName", {
-			_locale: i18n.getLocale(),
-		}),
-		companyId: i18n.__("SimpleSchema.labels.Salaries.companyId", {
-			_locale: i18n.getLocale(),
-		}),
-		jobTitle: i18n.__("SimpleSchema.labels.Salaries.jobTitle", {
-			_locale: i18n.getLocale(),
-		}),
-		incomeType: i18n.__("SimpleSchema.labels.Salaries.incomeType", {
-			_locale: i18n.getLocale(),
-		}),
-		incomeAmount: i18n.__("SimpleSchema.labels.Salaries.incomeAmount", {
-			_locale: i18n.getLocale(),
-		}),
-		gender: i18n.__("SimpleSchema.labels.Salaries.gender", {
-			_locale: i18n.getLocale(),
-		}),
-		datePosted: i18n.__("SimpleSchema.labels.Salaries.datePosted", {
-			_locale: i18n.getLocale(),
-		}),
-	};
-};
-
-const salaryErrors = function(locale) {
-	return {
-		noCompanyWithThatName: i18n.__("common.forms.companyNotFound", {
-			_locale: locale,
-		}),
-		sessionError: i18n.__("SimpleSchema.custom.sessionError", {
-			_locale: locale,
-		}),
-	};
-};
-
-const englishSalaries = salaryErrors("en");
-if (Meteor.isClient && Meteor.isProduction) {
-	i18n.loadLocale("es");
-}
-const spanishSalaries = salaryErrors("es");
-
-Salaries.schema.labels(salaryLabels());
-
-Salaries.schema.messageBox.messages({
-	en: englishSalaries,
-	es: spanishSalaries,
-});
-
-i18n.onChangeLocale(function(newLocale) {
-	if (Meteor.isDevelopment) console.log("SALARIES: " + newLocale);
-	Salaries.schema.messageBox.setLanguage(newLocale);
-	Salaries.schema.labels(salaryLabels());
-});
-
 Salaries.attachSchema(Salaries.schema, { replace: true });
 
 Salaries.deny({
